requirements.txt
requests==2.31.0
pre-commit==3.3.3
python-dotenv==1.0.0